Change protection to JSON only so as to support AoT/trimming and use of serialization options from MVC (#242)
Using JSON instead of the type descriptors helps avoid losing sub-second precision for the `DateTime` and `DateTimeOffset` types.

With this PR, `Protect(...)` and `Unprotect(...)` methods allow the passing of `JsonSerializerOptions` where trimming or AoT is not in use. These options are optional and when not supplied, the default JSON options configured for MVC are used. In trimming and AoT scenarios, these methods have overloads that accept `JsonTypeInfo<T>`.

As a side-effect, any pre-protected tokens may fail to unprotect if `options.UseConversionInsteadOfJson` had not been set to true. The workaround for this is to create a custom implementation of `ITokenProtector<T>` to handle the cases. For tokens with a short lifetime, this may not be an issue. The protector purpose has changed from `Tingle.Tokens.v2019-12-17` to `Tingle.AspNetCore.Tokens.v2024-05-05` to ensure one can find away support older scenarios