Upgrade `Tingle.AspNetCore.OpenApi` to .NET 10 Preview (#409)
This allows for use with the changes that are currently coming in such as XML comments. Consequently, adding extra tags is not longer supported but that can be sorted using `Microsoft.AspNetCore.Http.TagsAttribute`, `Microsoft.AspNetCore.Http.Metadata.ITagsMetadata`, or `endpoint.WithTags(...)`.

Other changes include support for OpenAPI 3.1, and handling of the document (e.g. tags now are referenced).

The `Microsoft.AspNetCore.OpenApi` package has been updated to a preview version.